fix key datatype in dist_sgd

+++ kge/util/dist_sgd.py
@@ -1,4 +1,5 @@
 import torch
+import numpy as np
 from torch.optim.optimizer import Optimizer, required
 
 class DistSGD(Optimizer):
@@ -123,7 +124,7 @@ class DistSGD(Optimizer):
                 #self.lapse_worker.push(self.lapse_indexes[i][indexes_to_push_mask], (-group['lr']*d_p).cpu().to_dense()[indexes_to_push_mask].numpy())
                 # TODO: this does not yet work with penalize
                 #  the mapping in penalize is still wrong -> take patricks freeze code as soon as ready
-                self.lapse_worker.push(self.local_to_lapse_mappers[i][indexes_to_push_mask], (-group['lr']*d_p).cpu().to_dense()[indexes_to_push_mask].numpy())
+                self.lapse_worker.push(self.local_to_lapse_mappers[i][indexes_to_push_mask].astype(np.uint64), (-group['lr']*d_p).cpu().to_dense()[indexes_to_push_mask].numpy())
                 #self.lapse_worker.push(self.lapse_indexes[i], (-group['lr']*d_p).cpu().to_dense().numpy())
                 p.add_(d_p, alpha=-group['lr'])
